Fixed debug report telemetry and sped up startup
Debug reports were not working correctly and the display of the games list from the quadstick was really slow.  When loading the file list, initially show only the CSV file names and put gathering the spreadsheet ID and name on a background thread that fills in the list later.
Fix an unhandled exception for when the OS will not allow reading the serial number from the flash drive to gather the firmware version.